% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BACps-internal}
\alias{BACps-internal}
\alias{logL}
\alias{MH.delta}
\alias{MH.alpha.s1_a}
\alias{MH.alpha.s1_b}
\alias{MH.delta.intercept}
\alias{countpattern}
\alias{logL_PS}
\alias{logL_Outcome}
\alias{update.params}
\alias{update.params_when_alpha_0}
\alias{updatemodel}
\title{Internal BACps functions}
\usage{
logL(N, y, U, alpha, delta)

MH.delta(
  N,
  y,
  U,
  delta0,
  alpha0,
  delta.mle.model.alpha0,
  delta.cov.model.alpha0
)

MH.alpha.s1_a(N, y, U, alpha0, alpha1, delta0, delta_t_1, q_delta_j)

MH.alpha.s1_b(N, y, U, alpha0, alpha1, delta0, delta_t_1, q_delta_j)

MH.delta.intercept(N, y, U, alpha0, delta0)

countpattern(x, matching = FALSE)

logL_PS(x, U, alphaX, gamma)

logL_Outcome(y, x, U, alphaX, gamma, Yparams)

\method{update}{params}(
  y,
  x,
  U,
  gamma_mle,
  gamma_cov,
  outcome_mle,
  outcome_cov,
  nparams,
  ngamma_t,
  params_t,
  alphaX,
  likhood_t
)

\method{update}{params_when_alpha_0}(y, x, U, nparams, ngamma_t, params_t, alphaX, likhood_t)

updatemodel(N, p, y, x, U, ngamma_t, params_t, alphaX_t, likhood_t, MX.nsims)
}
\description{
Internal BACps helper functions for the 1st Stage
}
\details{
These functions are not intended for use by users.
}
