% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACps-package.R
\docType{package}
\name{BACps-package}
\alias{BACps}
\alias{BACps-package}
\title{BACps: Bayesian Adjustment for Confouding in Bayesian Propensity Score}
\description{
Two stage procedure based on three models: prognostic score model,  propensity score model outcome model. The goal from the first stage is to obtain a posterior distribution of the model indicator of the propensity score to be used as an informative prior in the second stage. The second stage computes the average causal effect (conditional) using the informative prior. The informative prior provides a chance to rule out instrumental variables from the PS model, assigning less prior probability to those models that include instrumental variables.
}
\author{
\strong{Maintainer}: Pablo Gonzalez Ginestet \email{pablo.gonzalezginestet@mail.mcgill.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
