% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacps.R
\name{bacps}
\alias{bacps}
\title{Bayesian Adjustment for Confounding}
\usage{
bacps(w, x, y, U, niter_1st, niter_2nd, burnin_1st, burnin_2nd)
}
\arguments{
\item{w}{a data.frame with at least the following variables:}

\item{x}{a matrix with the exposure (binary) variable}

\item{y}{a matrix with the response (binary) variable}

\item{U}{a data frame with the covariates}

\item{niter_1st}{iteration MCMC for the 1st stage}

\item{niter_2nd}{iteration MCMC for the 2nd stage}

\item{burnin_1st}{burn-in MCMC for the 2nd stage}

\item{burnin_2nd}{burn-in MCMC for the 2nd stage}
}
\value{
a list
}
\description{
Main Algorithm that performs Bayesian Adjusting for confounding carrying out 1st and 2nd stage.
}
