% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulated data}
\usage{
sim_data(N, beta)
}
\arguments{
\item{N}{sample size}

\item{beta}{exposure marginal effect on outcome}
}
\value{
a list with outcome y,exposure x and covariates U
}
\description{
Function that generates simulated data with five covariates: three confounders,one instrumental variable and one predictor of the outcome (scenario from the paper of Wang)
}
